requirements.txt
telethon>=1.36.0
python-dotenv>=1.0.0
Pillow>=10.0.0
cryptg>=0.4.0
aiofiles>=23.0.0
